requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
beautifulsoup4
pdfkit
matplotlib
wordcloud
pandas
notebook
